♻️ Refactor: Extract NL2Agent MCP tools
